add eslint rules, add storybook setup, add @flow rules
