+++ client/.storybook/config.js
@@ -1,0 +1,9 @@
+/* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
+
+import { configure } from '@storybook/react';
+
+function loadStories() {
+  require('../app/stories/index');
+}
+
+configure(loadStories, module);
