--- client/app/bundles/HelloWorld/components/__tests__/Chart.spec.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import Chart from '../Chart';
-import { render } from 'enzyme';
-
-describe('Chart', () => {
-    it("renders a Chart", () => {
-        expect(() => {
-            render(<Chart />);
-        }).not.toThrow();
-    });
-});
