+++ client/app/bundles/momentDashboards/startup/registration.jsx
@@ -4,5 +4,5 @@ import Chart from '../components/Chart';
 
 // This is how react_on_rails can see the HelloWorld in the browser.
 ReactOnRails.register({
-    Chart,
+  Chart,
 });
