+++ client/webpack.config.js
@@ -21,7 +21,7 @@ const config = {
       'es5-shim/es5-shim',
       'es5-shim/es5-sham',
       'babel-polyfill',
-      './app/bundles/HelloWorld/startup/registration',
+      './app/bundles/momentDashboards/startup/registration',
     ],
   },
